Move get_credentials to authentication module.
